requirements.txt
robotframework
requests
robotframework-requests
robotframework-sudslibrary
robotframework-selenium2library